+++ CHANGELOG.md
@@ -5,8 +5,11 @@ This package was formerly known as `fmutool`.
 * ADDED: `fmucontainer` support of Terminals
 * ADDED: support of Clock and Binary types
 * ADDED: `fmucontainer` support of Event-Mode
-* ADDED: `fmucontainer` support of LS-BUS
+* ADDED: `fmucontainer` support of LS-BUS. See [ls-bus.md](doc/ls-bus.md)
+* ADDED: `fmucontainer` support of `-datalog` option to log values at each step. See [datalog.md](doc/datalog.md)
 * FIXED: `fmucontainer` log correctly for FMU3 embedded inside FMU2 container
+* FIXED: `fmucontainer` correctly set start values
+* FIXED: `fmucontainer` correctly handle tolerance. See [issue 35](https://github.com/grouperenault/fmu_manipulation_toolbox/issues/35)
 
 # Version 1.9.1.2
 * FIXED: Avoid usage of NULL pointer for category in logs. Despite this is legal,